feat(Component): Make error button a component
